fixed incorrect usage instructions
 ## Added

- pypi readme
- installation instructions in more detail for cases where an updated
version is not found

 ## Fixed

- incorrect usage instructions. i3 needs a full path to the
polybar-reload in the case of zsh users. i3 uses sh to execute commands
following the exec directive. zsh users usually don't add ~/.local/bin
to PATH in bashrc, which is what is sourced for sh. So instead of going
into details for different shell users in the readme i've opted for
people to just use a full path. works with all shells and the install
directory is always ~/.local/bin